requirements.txt
amqp==5.0.1
attrs==20.2.0
celery==5.0.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-repl==0.1.6
connexion[swagger-ui]==2.7.0
distlib==0.3.1
filelock==3.0.12
Flask==1.1.2
Flask-Environments==0.1
Flask-Login==0.5.0
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
Jinja2==2.11.2
kombu==5.0.2
Mako==1.1.3
MarkupSafe==1.1.1
pluggy==0.13.1
pytz==2020.1
PyYAML==5.3.1
redis==3.5.3
flask-redis==0.4.0
requests==2.24.0
six==1.15.0
SQLAlchemy==1.3.20
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.1
ujson==1.35
urllib3==1.25.11
Werkzeug==1.0.1
WTForms==2.3.3
timeago==1.0.14
dnspython==2.0.0
email-validator==1.1.1
importlib-resources==3.3.0
flask-rabmq==0.0.18
Pillow==8.4.0
